ai-engine/requirements.txt
requests
transformers
torch
accelerate
sentence-transformers
faiss-cpu
scikit-learn